src/tbd/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
streamlit==1.36.0
pymupdf==1.24.9
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.2
pdfminer.six==20240706
pytesseract==0.3.13
